Docling_Agents_PDFs/requirements.txt
requests
ipykernel
python-dotenv
openai
pydantic
docling
lancedb
streamlit
google-genai
google-generativeai